dragonfly_display/_extend_dragonfly.py
# coding=utf-8
# import the core dragonfly modules
from dragonfly.model import Model

# import the extension functions
from .model import model_to_vis_set

# inject the methods onto the classes
Model.to_vis_set = model_to_vis_set
